requirements.txt
flask
requests
Pillow
numpy
tensorflow
opencv-python
scikit-learn
gunicorn
